refactor: 각 컴포넌트에서 호출할 수 있는 함수 및 훅들을 이동시킴
