+++ src/pages/post-manager/ui/dialogs/EditPostDialog.tsx
@@ -1,3 +1,4 @@
+import { useUpdatePostMutation } from "@entities/post/api/mutations.ts"
 import { Post } from "@entities/post/model/types.ts"
 import {
   Button,
@@ -15,15 +16,27 @@ interface IProps {
   setShowEditDialog: React.Dispatch<React.SetStateAction<boolean>>
   selectedPost: Post | null
   setSelectedPost: React.Dispatch<React.SetStateAction<Post | null>>
-  updatePost: () => Promise<void>
 }
 export default function EditPostDialog({
   showEditDialog,
   setShowEditDialog,
   selectedPost,
   setSelectedPost,
-  updatePost,
+  limit,
+  skip,
 }: IProps) {
+  const updatePostMutation = useUpdatePostMutation({ limit, skip })
+
+  // 게시물 업데이트
+  const updatePost = async () => {
+    try {
+      await updatePostMutation.mutateAsync({ post: selectedPost })
+      setShowEditDialog(false)
+    } catch (error) {
+      console.error("게시물 업데이트 오류:", error)
+    }
+  }
+
   return (
     <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
       <DialogContent>
@@ -34,13 +47,19 @@ export default function EditPostDialog({
           <Input
             placeholder="제목"
             value={selectedPost?.title || ""}
-            onChange={(e) => setSelectedPost({ ...selectedPost, title: e.target.value })}
+            onChange={(e) => {
+              if (!selectedPost) return
+              setSelectedPost({ ...selectedPost, title: e.target.value })
+            }}
           />
           <Textarea
             rows={15}
             placeholder="내용"
             value={selectedPost?.body || ""}
-            onChange={(e) => setSelectedPost({ ...selectedPost, body: e.target.value })}
+            onChange={(e) => {
+              if (!selectedPost) return
+              setSelectedPost({ ...selectedPost, body: e.target.value })
+            }}
           />
           <Button onClick={updatePost}>게시물 업데이트</Button>
         </div>
